requirements.txt
aiogram==3.13.1
fastapi==0.115.4
uvicorn==0.32.0
SQLAlchemy==2.0.36
alembic==1.13.2
python-dotenv==1.0.1
aiofiles==24.1.0
pymysql==1.1.1
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.9.2
Babel==2.16.0
pytz==2024.2
rarfile==4.1

